Organize linalg dispatches according to library
