+++ pytensor/link/jax/dispatch/linalg/__init__.py
@@ -1,0 +1,8 @@
+from pytensor.link.jax.dispatch.linalg import (
+    constructors,
+    decomposition,
+    inverse,
+    products,
+    solve,
+    summary,
+)
