+++ pytensor/link/mlx/dispatch/__init__.py
@@ -16,6 +16,5 @@ import pytensor.link.mlx.dispatch.blockwise
 import pytensor.link.mlx.dispatch.extra_ops
 import pytensor.link.mlx.dispatch.pad
 import pytensor.link.mlx.dispatch.sort
-import pytensor.link.mlx.dispatch.slinalg
-import pytensor.link.mlx.dispatch.nlinalg
+import pytensor.link.mlx.dispatch.linalg
 # isort: on
