+++ pytensor/link/mlx/dispatch/linalg/__init__.py
@@ -1,0 +1,6 @@
+from pytensor.link.mlx.dispatch.linalg import (
+    decomposition,
+    inverse,
+    products,
+    solve,
+)
