+++ pytensor/link/mlx/dispatch/linalg/solve.py
@@ -3,23 +3,8 @@ import warnings
 import mlx.core as mx
 
 from pytensor.link.mlx.dispatch.basic import mlx_funcify
-from pytensor.tensor._linalg.decomposition.cholesky import Cholesky
-from pytensor.tensor._linalg.decomposition.lu import LU
 from pytensor.tensor._linalg.solve.general import Solve
 from pytensor.tensor._linalg.solve.triangular import SolveTriangular
-
-
-@mlx_funcify.register(Cholesky)
-def mlx_funcify_Cholesky(op, node, **kwargs):
-    lower = op.lower
-    a_dtype = getattr(mx, node.inputs[0].dtype)
-
-    def cholesky(a):
-        return mx.linalg.cholesky(
-            a.astype(dtype=a_dtype, stream=mx.cpu), upper=not lower, stream=mx.cpu
-        )
-
-    return cholesky
 
 
 @mlx_funcify.register(Solve)
@@ -35,7 +20,6 @@ def mlx_funcify_Solve(op, node, **kwargs):
         )
 
     def solve(a, b):
-        # MLX only supports solve on CPU
         return mx.linalg.solve(
             a.astype(stream=mx.cpu, dtype=a_dtype),
             b.astype(stream=mx.cpu, dtype=b_dtype),
@@ -56,32 +40,7 @@ def mlx_funcify_SolveTriangular(op, node, **kwargs):
             A.astype(stream=mx.cpu, dtype=A_dtype),
             b.astype(stream=mx.cpu, dtype=b_dtype),
             upper=not lower,
-            stream=mx.cpu,  # MLX only supports solve_triangular on CPU
+            stream=mx.cpu,
         )
 
     return solve_triangular
-
-
-@mlx_funcify.register(LU)
-def mlx_funcify_LU(op, node, **kwargs):
-    permute_l = op.permute_l
-    A_dtype = getattr(mx, node.inputs[0].dtype)
-    p_indices = op.p_indices
-
-    if permute_l:
-        raise ValueError("permute_l=True is not supported in the mlx backend.")
-    if not p_indices:
-        raise ValueError("p_indices=False is not supported in the mlx backend.")
-
-    def lu(a):
-        p_idx, L, U = mx.linalg.lu(
-            a.astype(dtype=A_dtype, stream=mx.cpu), stream=mx.cpu
-        )
-
-        return (
-            p_idx.astype(mx.int32, stream=mx.cpu),
-            L,
-            U,
-        )
-
-    return lu
