+++ pytensor/link/pytorch/dispatch/__init__.py
@@ -7,8 +7,7 @@ import pytensor.link.pytorch.dispatch.scalar
 import pytensor.link.pytorch.dispatch.elemwise
 import pytensor.link.pytorch.dispatch.math
 import pytensor.link.pytorch.dispatch.extra_ops
-import pytensor.link.pytorch.dispatch.nlinalg
-import pytensor.link.pytorch.dispatch.slinalg
+import pytensor.link.pytorch.dispatch.linalg
 import pytensor.link.pytorch.dispatch.shape
 import pytensor.link.pytorch.dispatch.sort
 import pytensor.link.pytorch.dispatch.subtensor
