+++ pytensor/link/pytorch/dispatch/linalg/__init__.py
@@ -1,0 +1,6 @@
+from pytensor.link.pytorch.dispatch.linalg import (
+    decomposition,
+    inverse,
+    products,
+    summary,
+)
